+++ src/schema/mod.rs
@@ -7,6 +7,7 @@ mod v6_to_v7;
 mod v7_to_v8;
 mod v8_to_v9;
 mod v9_to_v10;
+mod v10_to_v11;
 
 pub use base::CREATE_SCHEMA_SQL;
 pub use v2_to_v3::UPGRADE_2_TO_3_SQL;
@@ -17,3 +18,4 @@ pub use v6_to_v7::UPGRADE_6_TO_7_SQL;
 pub use v7_to_v8::UPGRADE_7_TO_8_SQL;
 pub use v8_to_v9::UPGRADE_8_TO_9_SQL;
 pub use v9_to_v10::UPGRADE_9_TO_10_SQL;
+pub use v10_to_v11::UPGRADE_10_TO_11_SQL;